backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase>=2.10.0
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
pymupdf==1.23.21
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
